AC:02 test pass by hunain

+++ app/src/main/java/parekhtdd/Item.java
@@ -1,12 +1,9 @@
 package parekhtdd;
-
-import java.math.BigDecimal;
-
 public class Item {
     String name;
-    BigDecimal price;
+    int price;
     Boolean onSale;
-    public Item(String name,BigDecimal price,Boolean onSale){
+    public Item(String name,int price,Boolean onSale){
         this.name = name;
         this.price = price;
         this.onSale = onSale;
